change timeout for receiving a message to depend on the peer-count
